Worker/requirements.txt
websockets
httpx
numpy
pillow
pytest
pytest-asyncio
uvicorn